requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.0
flake8==6.1.0
idna==3.4
mccabe==0.7.0
pyaes==1.6.1
pycodestyle==2.11.1
pyflakes==3.1.0
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.0.6
